zdm3o1-parallel-image-processing-pipeline-with-memory-safe-batch-operations/repository_before/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pillow==10.2.0
redis==5.0.1
python-multipart==0.0.6
